refactor(errors): move Wrap and Code into utils/errors package
Keep hypervisor focused on connect APIs; place Wrap, Code, AsLibvirtError,
and ToLegacy in utils/errors for shared use without stdlib name clash via alias.
